pyproject.toml
[tool.poetry]
name = "blanco-bot"
version = "1.0.0"
description = "Simple, no-frills Discord music bot"
authors = ["Jared Dantis <jareddantis@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "bot" }]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp-jinja2 = "^1.6"
aiohttp-session = "^2.12.0"
cryptography = "^42.0.5"
mafic = "^2.10.0"
nextcord = "^2.6.0"
pylast = "^5.2.0"
pyyaml = "^6.0.1"
ratelimit = "^2.2.1"
redis = "^5.0.3"
requests = "^2.31.0"
sentry-sdk = "^1.44.0"
spotipy = "^2.23.0"
tenacity = "^8.2.3"
thefuzz = "^0.22.1"
validators = "^0.24.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
pre-commit = "^3.7.0"
ruff = "^0.3.4"
watchfiles = "^0.21.0"

[tool.ruff]
indent-width = 2

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "PL"]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
